Move calling of jenkins steps to step itself
