+++ vars/piperPipelineStageMavenStaticCodeChecks.groovy
@@ -27,33 +27,6 @@ void call(Map parameters = [:]) {
     final script = checkScript(this, parameters) ?: null
 
     piperStageWrapper(stageName: stageName, script: script) {
-        try {
-            mavenExecuteStaticCodeChecks(script: script)
-        } catch (Exception exception) {
-            error("Maven Static Code Checks execution failed. Please examine the reports which are also available in the Jenkins user interface.")
-        }
-        finally {
-            showIssues(script, stageName)
-        }
-    }
-}
-
-private showIssues(Script script, String stageName) {
-    Map configuration = ConfigurationLoader.stageConfiguration(script, stageName)
-    // the checks are executed by default, even if they are not configured. They aren't executed only in case they are turned off with `false`
-    if (configuration.mavenExecuteStaticCodeChecks?.spotBugs == null || configuration.mavenExecuteStaticCodeChecks?.spotBugs == true) {
-        recordIssues(blameDisabled: true,
-            enabledForFailure: true,
-            aggregatingResults: false,
-            tool: spotBugs(pattern: '**/target/spotbugsXml.xml'))
-
-        ReportAggregator.instance.reportStaticCodeExecution(QualityCheck.FindbugsCheck)
-    }
-    if (configuration.mavenExecuteStaticCodeChecks?.pmd == null || configuration.mavenExecuteStaticCodeChecks?.pmd == true) {
-        recordIssues(blameDisabled: true,
-            enabledForFailure: true,
-            aggregatingResults: false,
-            tool: pmdParser(pattern: '**/target/pmd.xml'))
-        ReportAggregator.instance.reportStaticCodeExecution(QualityCheck.PmdCheck)
+        mavenExecuteStaticCodeChecks(script: script)
     }
 }
